Fix time sync wait on startup
We require time sync in order for the application to work
correctly. time-sync.target does not wait for the time to
be actually correct, so we use the systemd-wait-time-sync.service

We remove the network-online dependency, as the time-sync wait
pretty much ensures the network is online when we start. And
if we don't need time sync because we have an RTC, then we also
don't need the network to be online to start